removed testdata class in to the individual package
